models/opensora_backend.py
"""
Open-Sora backend for text+image to video generation.

Installation:
1. Clone Open-Sora repo: git clone https://github.com/hpcaitech/Open-Sora.git
2. Follow their installation instructions
3. Download weights to ./weights/opensora/

Note: This is a simplified wrapper. Open-Sora installation can be complex.
Adjust imports and paths based on your Open-Sora version.
"""
import logging
import time
from pathlib import Path
from typing import Optional, Dict
import torch
import numpy as np
from PIL import Image

logger = logging.getLogger("video_service")


class BackendNotAvailable(Exception):
    """Raised when backend cannot be initialized."""
    pass


class OpenSoraBackend:
    """
    Open-Sora backend wrapper for video generation.

    This implementation provides a minimal interface to Open-Sora.
    You may need to adjust based on your Open-Sora version.
    """

    def __init__(
        self,
        weights_dir: Path,
        device: str = "cuda",
        low_vram: bool = False
    ):
        """
        Initialize Open-Sora backend.

        Args:
            weights_dir: Path to weights directory
            device: torch device ('cuda' or 'cpu')
            low_vram: Enable low-VRAM optimizations
        """
        self.weights_dir = Path(weights_dir) / "opensora"
        self.device = device
        self.low_vram = low_vram
        self.model = None
        self.vae = None
        self.text_encoder = None

        logger.info(f"Initializing Open-Sora backend on {device}")

        if not torch.cuda.is_available() and device == "cuda":
            raise BackendNotAvailable("CUDA not available")

        self._load_models()

    def _load_models(self):
        """
        Load Open-Sora models.

        NOTE: This is a placeholder implementation. Open-Sora's API may vary.
        You'll need to adapt this based on the actual Open-Sora installation.
        """
        try:
            # Attempt to import Open-Sora modules
            # This assumes Open-Sora is installed and in PYTHONPATH
            # Adjust imports based on your Open-Sora version

            # Example structure (may need modification):
            # from opensora.models import STDiT, VAE
            # from opensora.pipelines import TextToVideoPipeline

            logger.warning(
                "Open-Sora backend is a placeholder implementation. "
                "Please integrate actual Open-Sora code based on your version."
            )

            # Check if weights exist
            if not self.weights_dir.exists():
                raise BackendNotAvailable(
                    f"Open-Sora weights not found at {self.weights_dir}. "
                    f"Please download weights and place them in this directory."
                )

            # Placeholder: In reality, you would load the actual models here
            # Example (adjust to actual Open-Sora API):
            #
            # self.model = STDiT.from_pretrained(
            #     self.weights_dir / "model",
            #     torch_dtype=torch.float16 if self.device == "cuda" else torch.float32
            # ).to(self.device)
            #
            # self.vae = VAE.from_pretrained(
            #     self.weights_dir / "vae"
            # ).to(self.device)
            #
            # self.text_encoder = ...
            #
            # if self.low_vram:
            #     self.model.enable_gradient_checkpointing()
            #     self.vae.enable_slicing()

            # For now, raise an exception prompting user to implement
            raise BackendNotAvailable(
                "Open-Sora backend requires manual integration. "
                "See opensora_backend.py for implementation guidance. "
                "Falling back to Mochi backend."
            )

        except ImportError as e:
            raise BackendNotAvailable(
                f"Open-Sora not installed or not in PYTHONPATH: {e}"
            )
        except Exception as e:
            raise BackendNotAvailable(f"Failed to load Open-Sora models: {e}")

    def generate_video(
        self,
        prompt: str,
        ref_image_path: Optional[str] = None,
        fps: int = 24,
        num_frames: int = 64,
        width: int = 720,
        height: int = 720,
        seed: Optional[int] = None
    ) -> Dict:
        """
        Generate video from text and optional reference image.

        Args:
            prompt: Text description
            ref_image_path: Optional path to reference image for I2V
            fps: Frames per second
            num_frames: Total number of frames
            width: Video width
            height: Video height
            seed: Random seed for reproducibility

        Returns:
            Dictionary with generation results
        """
        start_time = time.time()

        if seed is not None:
            torch.manual_seed(seed)
            np.random.seed(seed)

        logger.info(
            f"Generating with Open-Sora: {num_frames} frames at {width}x{height}, "
            f"{fps} fps, seed={seed}"
        )

        try:
            # Load reference image if provided
            ref_image = None
            if ref_image_path:
                ref_image = Image.open(ref_image_path).convert("RGB")
                ref_image = ref_image.resize((width, height), Image.LANCZOS)
                logger.info(f"Using reference image: {ref_image_path}")

            # Placeholder for actual generation
            # In a real implementation, you would:
            # 1. Encode the text prompt
            # 2. Optionally encode the reference image
            # 3. Run the diffusion model to generate latents
            # 4. Decode latents to video frames using VAE
            # 5. Save as video file

            # Example (pseudo-code):
            # with torch.no_grad():
            #     text_embeddings = self.text_encoder(prompt)
            #
            #     if ref_image:
            #         # Image-to-video mode
            #         image_embeddings = self.vae.encode(ref_image)
            #         latents = self.model(
            #             text_embeddings,
            #             image_cond=image_embeddings,
            #             num_frames=num_frames,
            #             height=height // 8,
            #             width=width // 8
            #         )
            #     else:
            #         # Text-to-video mode
            #         latents = self.model(
            #             text_embeddings,
            #             num_frames=num_frames,
            #             height=height // 8,
            #             width=width // 8
            #         )
            #
            #     frames = self.vae.decode(latents)
            #
            #     # Convert to video and save
            #     video_path = save_video(frames, fps, output_path)

            raise NotImplementedError(
                "Open-Sora backend generation not implemented. "
                "Please integrate with actual Open-Sora code."
            )

        except Exception as e:
            logger.error(f"Open-Sora generation failed: {e}")
            raise

        elapsed = time.time() - start_time
        logger.info(f"Generation completed in {elapsed:.2f}s")

        return {
            "mp4_path": str(video_path),
            "frames": num_frames,
            "fps": fps,
            "width": width,
            "height": height,
            "seed": seed,
            "backend": "opensora",
            "elapsed_time": round(elapsed, 2)
        }

    def __del__(self):
        """Cleanup GPU memory on deletion."""
        if torch.cuda.is_available():
            torch.cuda.empty_cache()
